final without download

+++ demo/routes/index.js
@@ -9,7 +9,7 @@ const path = require("path");
 const { Wallets, Gateway } = require('fabric-network');
 const FabricCAServices = require('fabric-ca-client');
 const fs = require("fs");
-const ccpPath = path.resolve('../config/connection-org1.json');
+const ccpPath = path.resolve('./config/connection-org1.json');
 const ccpJSON = fs.readFileSync(ccpPath, 'utf8');
 const ccp = JSON.parse(ccpJSON);
 
@@ -17,7 +17,8 @@ const cors = require('cors');
 router.use(cors({}));
 
 
-router.post("/", async function (req, res, next) {
+//인증서 등록 및 발급
+router.post("/tlsca/ca/approval", async function (req, res, next) {
 
   //let client = new Client();
 
@@ -44,11 +45,19 @@ router.post("/", async function (req, res, next) {
   //   });
   // }
 
-  //1. admin 인증서가 발행되어 있지 않으면 인증서 발행
-  //2. 발행된 인증서를 사용하여 유저 생성
-
 
   try{
+
+    let token = req.headers.token;
+
+    const result = await jwt.verify(token); // verify 함수 호출
+    if(result < 0){
+      res.status(401).json({
+        "result": false,
+        "message" : "invalid token"
+      });
+      return;
+    }
 
     const caInfo_tls = ccp.certificateAuthorities['ca.tls.hypercerts.com'];
     const caInfo_org1 = ccp.certificateAuthorities['ca.org1.hypercerts.com'];
@@ -70,64 +79,80 @@ router.post("/", async function (req, res, next) {
           if(rows[i].net_group === "tls"){
             console.log("tls");
 
-            // const enrollment = await ca_tls.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
-            // const x509Identity = {
-            //   credentials: {
-            //     certificate: enrollment.certificate,
-            //   },
-            // };
-            //
-            // await wallet.put(rows[i].id, x509Identity);
+            const enrollment = await ca_tls.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
+            const x509Identity = {
+              credentials: {
+                certificate: enrollment.certificate,
+                privateKey: enrollment.key.toBytes(),
+              },
+              roles: rows[i].role,
+              mspId: 'Org1MSP',
+              type: 'X.509',
+            };
+
+            await wallet.put(rows[i].id, x509Identity);
 
           }else if(rows[i].net_group === "org1"){
 
             console.log("org1");
 
-            // const enrollment = await ca_org1.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
-            // const x509Identity = {
-            //   credentials: {
-            //     certificate: enrollment.certificate,
-            //   },
-            // };
-            //
-            // await wallet.put(rows[i].id, x509Identity);
+            const enrollment = await ca_org1.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
+            const x509Identity = {
+              credentials: {
+                certificate: enrollment.certificate,
+                privateKey: enrollment.key.toBytes(),
+              },
+              roles: rows[i].role,
+              mspId: 'Org1MSP',
+              type: 'X.509',
+            };
+
+            await wallet.put(rows[i].id, x509Identity);
 
           }else if(rows[i].net_group === "org2"){
 
             console.log("org2");
 
-            // const userExists = await wallet.get(rows[i].id);
-            // if(!userExists){
-            //
-            //   const enrollment = await ca_org2.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
-            //
-            //   const x509Identity = {
-            //     credentials: {
-            //       certificate: enrollment.certificate,
-            //     },
-            //   };
-            //
-            //   await wallet.put(rows[i].id, x509Identity);
-            // }
+            const userExists = await wallet.get(rows[i].id);
+            if(!userExists){
+
+              const enrollment = await ca_org2.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
+
+              const x509Identity = {
+                credentials: {
+                  certificate: enrollment.certificate,
+                  privateKey: enrollment.key.toBytes(),
+                },
+                roles: rows[i].role,
+                mspId: 'Org1MSP',
+                type: 'X.509',
+              };
+
+              await wallet.put(rows[i].id, x509Identity);
+            }
 
           }else{
 
             console.log("orderer");
 
-            // const userExists = await wallet.get(rows[i].id);
-            // if(!userExists){
-            //
-            //   const enrollment = await ca_orderer.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
-            //
-            //   const x509Identity = {
-            //     credentials: {
-            //       certificate: enrollment.certificate,
-            //     },
-            //   };
-            //
-            //   await wallet.put(rows[i].id, x509Identity);
-            // }
+            const userExists = await wallet.get(rows[i].id);
+            if(!userExists){
 
+              const enrollment = await ca_orderer.enroll({ enrollmentID: rows[i].id, enrollmentSecret: rows[i].id + "pw"});
+
+              const x509Identity = {
+                credentials: {
+                  certificate: enrollment.certificate,
+                  privateKey: enrollment.key.toBytes(),
+                },
+                roles: rows[i].role,
+                mspId: 'Org1MSP',
+                type: 'X.509',
+              };
+
+              await wallet.put(rows[i].id, x509Identity);
+
+            }
 
           }
 
@@ -137,9 +162,9 @@ router.post("/", async function (req, res, next) {
           "result": true,
         });
       } else {
-        res.status(200).json({
+        res.status(400).json({
           "result": false,
-          "message": "not match"
+          "message": "db error"
         });
       }
 
@@ -147,7 +172,7 @@ router.post("/", async function (req, res, next) {
 
   }catch (err){
     console.log(err);
-    res.status(200).json({
+    res.status(400).json({
       "result": false,
       "message": "error"
     });
@@ -168,10 +193,21 @@ router.post("/", async function (req, res, next) {
 });
 
 //인증서 생성
-router.post('/tlsca/ca/create', function(req, res) {
+router.post('/tlsca/ca/create', async function(req, res) {
 
   console.log("/tlsca/ca/create");
   console.log(req.body);
+
+  let token = req.headers.token;
+
+  const result = await jwt.verify(token); // verify 함수 호출
+  if(result < 0){
+    res.status(401).json({
+      "result": false,
+      "message" : "invalid token"
+    });
+    return;
+  }
 
   let user_id = req.body.id; //아이디
   let group = req.body.group; //소속
@@ -188,7 +224,7 @@ router.post('/tlsca/ca/create', function(req, res) {
         });
     } else {
       console.log('err : ' + err);
-      res.status(200).json({
+      res.status(400).json({
         "result": false,
         "message": "error"
       });
@@ -199,9 +235,20 @@ router.post('/tlsca/ca/create', function(req, res) {
 });
 
 //인증서 리스트 조회
-router.get('/tlsca/ca/list', function(req, res) {
+router.get('/tlsca/ca/list', async function(req, res) {
 
   console.log("/tlsca/ca/list");
+  let token = req.headers.token;
+
+    const result = await jwt.verify(token); // verify 함수 호출
+    if(result < 0){
+      res.status(401).json({
+        "result": false,
+        "message" : "invalid token"
+      });
+      return;
+    }
+
 
   db.query('select * from ca', async function (err, rows, field) {
     if (!err) {
@@ -210,7 +257,7 @@ router.get('/tlsca/ca/list', function(req, res) {
           "list" : rows
         });
       } else {
-        res.status(200).json({
+        res.status(400).json({
           "result": false,
           "message": "not match"
         });
@@ -221,9 +268,20 @@ router.get('/tlsca/ca/list', function(req, res) {
 });
 
 //인증서 삭제
-router.delete('/tlsca/ca/delete/:pk', function(req, res) {
+router.delete('/tlsca/ca/delete/:pk', async function(req, res) {
 
   console.log(req.params.pk);
+
+  let token = req.headers.token;
+
+  const result = await jwt.verify(token); // verify 함수 호출
+  if(result < 0){
+    res.status(401).json({
+      "result": false,
+      "message" : "invalid token"
+    });
+    return;
+  }
 
   var number = req.params.pk;
   let query = "delete from ca where pk = "+number;
@@ -233,7 +291,7 @@ router.delete('/tlsca/ca/delete/:pk', function(req, res) {
         "result": true,
       });
     } else {
-      res.status(200).json({
+      res.status(400).json({
         "result": false,
         "message": "error"
       });
@@ -244,10 +302,19 @@ router.delete('/tlsca/ca/delete/:pk', function(req, res) {
 });
 
 //인증서 업데이트 및 수정
-router.post('/tlsca/ca/update/:pk', function(req, res) {
+router.post('/tlsca/ca/update/:pk', async function(req, res) {
 
-  console.log(req.params.pk);
-  console.log(req.body);
+  let token = req.headers.token;
+
+  const result = await jwt.verify(token); // verify 함수 호출
+  if(result < 0){
+    res.status(401).json({
+      "result": false,
+      "message" : "invalid token"
+    });
+    return;
+  }
+
   var number = req.params.pk;
 
   let query = "update ca set id='"+req.body.id+"', net_group='"+req.body.group +"', role='"+req.body.role +"' where pk="+number;
@@ -258,7 +325,7 @@ router.post('/tlsca/ca/update/:pk', function(req, res) {
         "result": true,
       });
     } else {
-      res.status(200).json({
+      res.status(400).json({
         "result": false,
         "message": "error"
       });
@@ -268,132 +335,112 @@ router.post('/tlsca/ca/update/:pk', function(req, res) {
   });
 });
 
+
+//인증서 다운로드
+router.get('/tlsca/ca/get/:id', async function(req, res) {
+
+  let token = req.headers.token;
+
+  const result = await jwt.verify(token); // verify 함수 호출
+  if (result < 0) {
+    res.status(401).json({
+      "result": false,
+      "message": "invalid token"
+    });
+    return;
+  }
+
+  console.log(req.params.id);
+  res.setHeader('Content-type', "text/html"); // 파일 형식 지정
+  res.download(`./wallet/${req.params.id}`, (err) => {
+    if (err) {
+      // 에러 처리
+      res.status(404).send('File not found');
+    } else {
+      console.log('File downloaded successfully');
+    }
+  });
+
+});
+
+//체인코드 연동 준비
+// router.get('/query', async function(req, res, next) {
+//
+//   const walletPath = path.join(process.cwd(), 'wallet');
+//   const wallet = await Wallets.newFileSystemWallet(walletPath);
+//   console.log(`wallet path : ${walletPath}`);
+//
+//   //const userExists = await wallet.get("user1");
+//   if(!userExists){
+//     console.log("not exist user1");
+//     res.status(200).json({
+//       "result" : false,
+//       "message" : "not exist user1"
+//
+//     });
+//   }
+//
+//   const gateway = new Gateway();
+//   await gateway.connect(ccp, {wallet, identify: 'user1', discovery: {enabled: false}});
+//
+//   const network = await gateway.getNetwork('mychannel');
+//
+//   const contract = network.getContract('fabcar');
+//
+//   const result = await contract.evaluateTransaction('queryAllCars');
+//
+//
+//   res.status(200).json({
+//     "result" : true,
+//     "message" : "not exist user1",
+//     "data" : result
+//   });
+//
+// });
+
+//관리자 로그인
+    router.post('/auth/ca/admin/login', function(req, res, next) {
+
+      let query = "select * from account where account='"+ req.body.id +"'";
+      db.query(query, async function (err, rows, field) {
+        if (!err) {
+          console.log(req.body.id);
+          if ((req.body.pw === rows[0].pw)) {
+            const token = await jwt.sign(1);
+
+            console.log("token : " + token);
+
+            res.status(200).json({
+              "result": true,
+              "token" : token
+            });
+          } else {
+            res.status(200).json({
+              "result": false,
+              "message": "not match"
+            });
+          }
+          // console.log(rows[0].id);
+
+        } else {
+          console.log('err : ' + err);
+          res.status(200).json({
+            "result": false,
+            "message": "error"
+          });
+        }
+      })
+    });
+
+//로그아웃
 router.get('/auth/ca/admin/logout', async function(req, res, next) {
 
   let token = req.headers.token;
 
   console.log(token);
 
-  res.status(200).json({
+  res.status(401).json({
     "result" : true,
   });
-
 });
-
-
-router.get('/check', async function(req, res, next) {
-
-  let token = req.headers.token;
-
-  console.log(token);
-
-  const result = await jwt.verify(token);
-
-  console.log(result);
-
-  res.status(200).json({
-    "result" : true,
-  });
-
-});
-
-router.get('/query', async function(req, res, next) {
-
-  const walletPath = path.join(process.cwd(), 'wallet');
-  const wallet = await Wallets.newFileSystemWallet(walletPath);
-  console.log(`wallet path : ${walletPath}`);
-
-  //const userExists = await wallet.get("user1");
-  if(!userExists){
-    console.log("not exist user1");
-    res.status(200).json({
-      "result" : false,
-      "message" : "not exist user1"
-    });
-  }
-
-  const gateway = new Gateway();
-  await gateway.connect(ccp, {wallet, identify: 'user1', discovery: {enabled: false}});
-
-  const network = await gateway.getNetwork('mychannel');
-
-  const contract = network.getContract('fabcar');
-
-  const result = await contract.evaluateTransaction('queryAllCars');
-
-
-  res.status(200).json({
-    "result" : true,
-    "message" : "not exist user1",
-    "data" : result
-  });
-
-
-  res.status(200).json({
-    "result" : true,
-  });
-
-});
-
-router.post('/auth/ca/admin/login', function(req, res, next) {
-
-  db.query('select * from account', async function (err, rows, field) {
-    if (!err) {
-      console.log(req.body.id);
-      if ((req.body.id === rows[0].id) && (req.body.pw === rows[0].pw)) {
-        const token = await jwt.sign(1);
-
-        console.log("token : " + token);
-
-        res.status(200).json({
-          "result": true,
-          "token" : token
-        });
-      } else {
-        res.status(200).json({
-          "result": false,
-          "message": "not match"
-        });
-      }
-      // console.log(rows[0].id);
-
-    } else {
-      console.log('err : ' + err);
-      res.status(200).json({
-        "result": false,
-        "message": "error"
-      });
-    }
-  })
-});
-
-router.get('/auth/ca/admin/login', function(req, res, next) {
-
-  db.query('select * from account', async function (err, rows, field) {
-    if (!err) {
-
-      console.log(req.body.id);
-
-      const token = await jwt.sign(1);
-
-      console.log("token : " + token);
-
-      res.status(200).json({
-        "result": true,
-        "data" : rows
-      });
-
-      // console.log(rows[0].id);
-
-    } else {
-      console.log('err : ' + err);
-      res.status(200).json({
-        "result": false,
-        "message": "error"
-      });
-    }
-  })
-});
-
-
 module.exports = router;
